Added Entity HP and Damage
HP has been implemented for Enemies and the Player. For a general aproach a general Entity Class has been created. Each enemy and the player will take this class as parent, so it inherits the base functionalities. A simple HP bar has been added as well.
Each implementation is very simplified and needs to be optimized